mysite/requirements.txt
bleach==2.1.4
certifi==2018.8.24
Django==2.0.8
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
Faker==0.8.8
html5lib==1.0.1
mixer==6.0.1
Pillow==5.2.0
pipenv==2018.7.1
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
virtualenv==16.0.0
virtualenv-clone==0.3.0
webencodings==0.5.1
